Update synaptic parameters for Sargent scaling.
As in Schwart2012, AMPA waveforms are scaled so that the peak value
matches the one reported in Sargent2005, and NMDA waveforms are scaled
so that when the Mg2+ block is completely removed they too have the
same peak amplitude ('phi' parameter defined in Schwartz2012 set to 1,
assumed to be a 'physiological' value). Keep in mind that the actual
value of peak gNMDAR will always be far less than gAMPAR because of
the block: at -40mV (spiking threshold), the actual conductance is
about 20% of the unblocked value.